+++ OOP/class_creature.py
@@ -1,0 +1,11 @@
+class Creature:
+
+    def __init__(self, name, age):
+        self.name = name
+        self.age = age
+
+    def get_info(self):
+        return f"Name: {self.name}, Age: {self.age}"
+
+    def speak(self):
+        return "I am a creature"
