+++ OOP/multi_inheritance/class_fish.py
@@ -1,0 +1,5 @@
+from class_prey import Prey
+from class_predator import Predator
+
+class Fish(Prey, Predator):
+    pass
