+++ OOP/multi_inheritance/class_predator.py
@@ -1,0 +1,5 @@
+class Predator:
+
+    def hunting(self):
+        print("Predator is hunting!")
+
