+++ OOP/multi_inheritance/class_prey.py
@@ -1,0 +1,4 @@
+class Prey:
+
+    def flee(self):
+        print("Prey is fleeing!")
