+++ OOP/multi_inheritance/class_rabbit.py
@@ -1,0 +1,4 @@
+from class_prey import Prey
+
+class Rabbit(Prey):
+    pass
