+++ OOP/multi_inheritance/class_wolf.py
@@ -1,0 +1,4 @@
+from class_predator import Predator
+
+class Wolf(Predator):
+    pass
