+++ OOP/multi_inheritance/main.py
@@ -1,0 +1,13 @@
+from class_fish import Fish
+from class_wolf import Wolf
+from class_rabbit import Rabbit
+
+
+rabbit = Rabbit()
+wolf = Wolf()
+fish = Fish()
+
+wolf.hunting()
+rabbit.flee()
+fish.hunting()
+fish.flee()
